:sparkles: feat: Added exceptions handlers

+++ src/main/java/com/lucasescol/webservicespringboot/services/UserService.java
@@ -4,14 +4,19 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.lucasescol.webservicespringboot.entities.User;
 import com.lucasescol.webservicespringboot.repositories.UserRepository;
+import com.lucasescol.webservicespringboot.services.exceptions.DatabaseException;
+import com.lucasescol.webservicespringboot.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
-    
+
     @Autowired
     private UserRepository repository;
 
@@ -21,7 +26,7 @@ public class UserService {
 
     public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insert(User obj) {
@@ -29,13 +34,25 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);			
+            } else {				
+                throw new ResourceNotFoundException(id);	
+            }
+        } catch (DataIntegrityViolationException e) {			
+            throw new DatabaseException(e.getMessage());		
+        }	
     }
 
     public User update(Long id, User obj) {
-        User entity = repository.getReferenceById(id);
-        updateData(entity, obj);
-        return repository.save(entity);
+        try {
+            User entity = repository.getReferenceById(id);
+            updateData(entity, obj);
+            return repository.save(entity);
+        } catch(EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User obj) {
